feat(interaction): share debug overlay across interactive sketches
Extracts the crosshair lines, pointer markers, finger chains, webcam
preview and per-source legend from the hardcoded interaction-test
sketch into a new utils/interaction/overlay.js module, then wires it
into splines-v1-interactive so the same Visualization and Vision -
Camera options actually do something in that sketch.

The spline demonstration overlay (raw polygon + point markers) now
ships ON by default in splines-v1-interactive to match v0, since the
render path already routed it through the shared _shared.js helpers.

Adds Show pointer markers / Show source legend toggles so each piece
of the debug overlay can be turned off independently.
